Give up on protecting against user-provided symlinks to arbitrary targets. (#10)
We just can't do it right. We can't forbid all symlinks, as symlinks are often
used by sysadmins to provide consistent aliases to mounted filesystems across a
HPC cluster; trying to distinguish between legal and illegal symlinks involves
some uncomfortable level of heuristics. For example, if the same filesystem is 
mounted elsewhere under a different name, all our heuristics break.

Better to assume that everything visible to the service account is fully
public, and to plan accordingly; this avoids making promises that we can't keep
whenever someone figures out a new symlink-based exploit in the future.