Periodically do manual checkpoints to sync the DB file with the WAL.
This is more efficient as it allows idle connections to stick around, reducing
the overhead of sporadic queries. For example, opening the first new connection
will cause creation of new WAL files (and then their subsequent deletion when
all connections are closed); this work is avoided by just keeping it open.