Simplified the parent-path-finding logic to be strictly lexical.
We no longer query the filesystem to check if each of the parent paths exist,
which improves efficiency and avoids unnecessary failures when a parent path
cannot be listed (e.g., due to execute-only permissions on a directory).
Checks for an existing directory have been moved into the list* functions
themselves, which is arguably a more appropriate place for them anyway.

This change means that a user can now use the /list/files endpoint to list
files through a symlinked directory inside a registered directory. Technically,
this is inconsistent with our policy of not traversing symlinked directories,
but if the user attempts to do so, that's their problem, not ours.