Use more conventional wildcards for pattern matching. (#16)
The usual ? and * wildcards are more intuitive. Also, there's no reason to
expose the user to the fact that we're using SQLite under the hood.

We rename the 'partial' property to 'is_pattern', which is more intuitive. The
same option is also used for pattern matching on paths for consistency.