Modified listMetadata to never return an error.
Any problems are reported in the failures. This occurs even if the supplied
path does not exist or is not a directory, in which case a failure is reported
and no metadata files are listed. The aim is to ensure that
addDirectoryContents does not quit early due to listing problems.

Also modified listFiles to shift directory verification into its associated
REST handler, to be consistent with the other uses of verifyDirectory.