Optionally restrict link whitelisting to certain users.
This aims to reduce the potential for DoS abuse when the average user just
creates a symlink to a large whitelisted directory, which would force SewerRat
to recurse into it. If traversal is only respected for symlinks created by
certain users, we can limit this behavior to people we trust.