Introduce within_path= option for listing registered directories.
This actually checks that the registered directory lies within the filter path
specified by within_path=. By comparison, path_prefix= just checks the prefix,
so if people don't add a trailing slash, they'll pick up directories with the
same name prefix, which is rarely desirable. (Conversely, if a trailing slash
is included, we won't pick up registered paths that are equal to the filter
path, as the former are not stored in the database with trailing slashes.)

The goal is to deprecate the path_prefix= option as people should really be
using within_path; it's unlikely they want to filter on the prefix of names.