Forbid all symbolic links unless they are owned by whitelisted users. (#29)
This closes a security hole in the previous whitelisting scheme whereby a user
could create a symbolic link to an approved directory, allow SewerRat to think
that it's okay, and then replace it with a link to another directory when
SewerRat gets around to actually reading the contents of the link.

Now, a symbolic link can only be owned by a trusted user, regardless of where
it points. This represents a more absolute trust boundary - if the user is
trusted, we can assume that they won't redirect the link to somewhere dangerous
(e.g., our SSH credentials), and we handle the link target like normal. If the
user is untrusted, we just refuse to traverse the symbolic link. 

This applies to both links inside the registered directory, and any symbolic
links in the path to the registered directory. Extension of this permission
scheme to the latter is necessary as an untrusted user could redirect their
symbolic link after registration, allowing them to pull out juicy secrets. 

We default the whitelist to 'root' to ensure that root-created symlinks (e.g.,
for aliasing mounts on complex filesystems) are automatically respected. We
also update the tests to include the current user and all owners of the
temporary directory in the whitelist by default, to avoid issues with CI.